# apps/docs/src/openapi/tradr-api.json
{
  "components": {
    "schemas": {
...
      "WidgetPlacement": {
        "description": "One widget on the dashboard grid. The grid is 12 columns wide and rows are a fixed height, so `x`/`w` are columns and `y`/`h` are rows.\n",
        "properties": {
          "config": {
            "description": "Optional per-widget settings. Serialises to at most 2048 bytes."
          },
          "h": {
            "maximum": 24,
            "minimum": 1,
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "enum": [
              "stats-summary",
              "open-positions",
              "performance-chart",
              "account-balances",
              "position-sizing",
              "equity-curve"
            ],
            "type": "string"
          },
          "w": {
            "description": "x + w must not exceed 12.",
            "maximum": 12,
            "minimum": 1,
            "type": "integer"
          },
          "x": {
            "maximum": 11,
            "minimum": 0,
            "type": "integer"
          },
          "y": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "type",
          "x",
          "y",
          "w",
          "h"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "HTTP API for Tradr — the open-source trading journal with an AI advisor.\n\nThis reference is generated from the `@swagger` JSDoc blocks that live next to each route in the API source, so it always matches the endpoints the app actually serves. All routes are mounted under `/api`. The same API backs both the hosted app (`api.tradr.cloud`) and a self-hosted instance (reached through the web container at `/api`). Most endpoints require an authenticated session cookie.",
    "title": "Tradr API",
    "version": "0.5.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/dashboard/layout": {
      "get": {
        "description": "Authed. Returns the user's widget placements on a 12-column grid, their theme, and when the layout was last saved. A user who has never customised the dashboard gets the default layout, not an empty one.\n",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "theme": {
                      "enum": [
                        "light",
                        "dark",
                        "system"
                      ],
                      "type": "string"
                    },
                    "updatedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "widgets": {
                      "items": {
                        "$ref": "#/components/schemas/WidgetPlacement"
                      },
                      "type": "array"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The layout."
          },
          "401": {
            "description": "No valid session."
          }
        },
        "summary": "Get the dashboard layout.",
        "tags": [
          "Dashboard"
        ]
      },
      "put": {
        "description": "Authed. Send `widgets`, `theme`, or both — a body with neither is a 400. `widgets` replaces the whole layout; it is not a patch. Placements are validated as a set: at most one widget of each type, no two overlapping, none extending past the 12-column grid, and none below its type's minimum size. When `theme` is sent the response also sets the `tradr_theme` cookie so the next page load paints without a flash.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "theme": {
                    "enum": [
                      "light",
                      "dark",
                      "system"
                    ],
                    "type": "string"
                  },
                  "widgets": {
                    "items": {
                      "$ref": "#/components/schemas/WidgetPlacement"
                    },
                    "maxItems": 6,
                    "type": "array"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The saved layout."
          },
          "400": {
            "description": "Validation error",
            "or a body containing neither widgets nor theme.": null
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE — the request body exceeds the limit."
          }
        },
        "summary": "Save the dashboard layout, the theme, or both.",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/dashboard/theme-cookie": {
      "post": {
        "description": "Authed. Sets `tradr_theme` from the stored preference without changing it. Used to restore the cookie on a new device or after it is cleared, so the first paint matches the saved theme. The request body must be empty.\n",
        "responses": {
          "204": {
            "description": "The cookie is set."
          },
          "400": {
            "description": "The request body was not empty."
          }
        },
        "summary": "Re-issue the theme cookie.",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/positions/{id}/fills": {
      "post": {
        "description": "Authed. A position is a sequence of fills, not a single trade: add an `entry` fill to open or scale in, an `exit` fill to scale out or close. The position's average price, open quantity, status, and realised P&L are all derived from its fills, so adding one recomputes them. Monetary and quantity fields are strings to preserve decimal precision.\n",
        "parameters": [
          {
            "description": "The position the fill belongs to.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "fees": {
                    "default": "0",
                    "description": "Non-negative decimal.",
                    "type": "string"
                  },
                  "filledAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "notes": {
                    "maxLength": 10000,
                    "nullable": true,
                    "type": "string"
                  },
                  "price": {
                    "description": "Non-negative decimal.",
                    "type": "string"
                  },
                  "quantity": {
                    "description": "Positive decimal.",
                    "type": "string"
                  },
                  "type": {
                    "enum": [
                      "entry",
                      "exit"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "type",
                  "price",
                  "quantity",
                  "filledAt"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The created fill."
          },
          "400": {
            "description": "Validation error: an exit quantity beyond the available entry quantity, or a fractional quantity on an options position."
          },
          "404": {
            "description": "No such position for this user."
          },
          "409": {
            "description": "The position is closed",
            "or an exit fill was sent to a position that is still a draft.": null
          }
        },
        "summary": "Add a fill to a position.",
        "tags": [
          "Positions"
        ]
      }
    },
    "/api/positions/{id}/fills/{fillId}": {
      "delete": {
        "description": "Authed. Removes the fill and recomputes the position from what is left. An open position must keep at least one entry fill, and its entry quantity must stay at or above its exit quantity, so a deletion that would break either rule is refused rather than applied.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "fillId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "404": {
            "description": "No such position or fill for this user."
          },
          "409": {
            "description": "The position is closed",
            "or removing it would leave exit fills unbacked.": null,
            "the fill is the last remaining entry": null
          }
        },
        "summary": "Delete a fill.",
        "tags": [
          "Positions"
        ]
      },
      "put": {
        "description": "Authed. Corrects a recorded fill and recomputes the position's derived figures. `type` is immutable — an entry cannot become an exit; delete the fill and add the right one instead. Every field is optional; only what you send changes.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "fillId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "fees": {
                    "description": "Non-negative decimal.",
                    "type": "string"
                  },
                  "filledAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "notes": {
                    "maxLength": 10000,
                    "nullable": true,
                    "type": "string"
                  },
                  "price": {
                    "description": "Non-negative decimal.",
                    "type": "string"
                  },
                  "quantity": {
                    "description": "Positive decimal.",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated fill."
          },
          "400": {
            "description": "Validation error: an exit quantity beyond the available entry quantity, or a fractional quantity on an options position."
          },
          "404": {
            "description": "No such position or fill for this user."
          },
          "409": {
            "description": "The change would leave the position inconsistent with its other fills."
          }
        },
        "summary": "Edit a fill.",
        "tags": [
          "Positions"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Hosted (tradr.cloud)",
      "url": "https://api.tradr.cloud"
    },
    {
      "description": "Local development (API PORT default)",
      "url": "http://localhost:3100"
    }
  ],
...
}